pyproject.toml
[project]
name = "inventorysynccli"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.24",
]
